Rename the config_wordalgebra module to globals
